# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IntelliRoute graph logistics API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  neo4j: string;
  /** @nullable */
  message?: string | null;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  Admin: 'Admin',
  Operations_Manager: 'Operations Manager',
  Delivery_Agent: 'Delivery Agent',
} as const;

export interface LoginInput {
  role: LoginInputRole;
}

export interface Session {
  role: string;
  userName: string;
}

export interface Customer {
  customerId: string;
  name: string;
  email: string;
  phone: string;
  status: string;
  /** @nullable */
  createdAt?: string | null;
  orderCount: number;
  shipmentCount: number;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 5 */
  phone: string;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: string;
}

export type CustomerDetailAddressesItem = { [key: string]: unknown };

export interface Order {
  orderId: string;
  customerId: string;
  amount: number;
  paymentType: string;
  priority: string;
  status: string;
  orderDate: string;
  itemCount?: number;
  /** @nullable */
  shipmentId?: string | null;
}

export interface Shipment {
  shipmentId: string;
  orderId: string;
  customerId: string;
  customerName: string;
  zone: string;
  agent: string;
  hub: string;
  /** @nullable */
  vehicle?: string | null;
  status: string;
  shipmentDate?: string;
  expectedDeliveryDate?: string;
  distanceKm: number;
  attemptCount: number;
  /** @nullable */
  failureProbability: number | null;
  /** @nullable */
  riskLevel?: string | null;
}

export type CustomerDetail = Customer & {
  addresses: CustomerDetailAddressesItem[];
  orders: Order[];
  shipments: Shipment[];
};

export interface OrderInput {
  customerId: string;
  amount: number;
  paymentType: string;
  priority: string;
  productName?: string;
  /** @minimum 1 */
  quantity?: number;
}

export type ShipmentDetailAddress = { [key: string]: unknown };

export type ShipmentDetailFailureReasonsItem = { [key: string]: unknown };

export interface DeliveryAttempt {
  attemptId: string;
  attemptNumber: number;
  status: string;
  attemptDate: string;
  remarks: string;
  /** @nullable */
  failureReason?: string | null;
}

export type ShipmentDetail = Shipment & {
  address: ShipmentDetailAddress;
  attempts: DeliveryAttempt[];
  failureReasons: ShipmentDetailFailureReasonsItem[];
};

export interface ShipmentInput {
  orderId: string;
  customerId: string;
  addressId: string;
  agentId: string;
  hubId: string;
  zoneId: string;
  /** @nullable */
  vehicleId?: string | null;
  shipmentDate: string;
  expectedDeliveryDate: string;
  distanceKm: number;
  deliveryTimeWindow?: string;
}

export interface ShipmentStatusInput {
  status: string;
}

export interface AttemptInput {
  status: string;
  /** @nullable */
  failureReasonId?: string | null;
  remarks: string;
}

export type GraphNodeProperties = { [key: string]: unknown };

export interface GraphNode {
  id: string;
  label: string;
  type: string;
  properties: GraphNodeProperties;
}

export interface GraphRelationship {
  source: string;
  target: string;
  type: string;
}

export interface GraphResponse {
  nodes: GraphNode[];
  relationships: GraphRelationship[];
}

export type DashboardSummaryMetrics = {[key: string]: number};

export type DashboardSummaryStatusDistributionItem = { [key: string]: unknown };

export type DashboardSummaryFailureReasonsItem = { [key: string]: unknown };

export type DashboardSummaryZonePerformanceItem = { [key: string]: unknown };

export type DashboardSummaryAgentPerformanceItem = { [key: string]: unknown };

export interface HighRiskShipment {
  shipmentId: string;
  customerName: string;
  zone: string;
  failureProbability: number;
  riskLevel: string;
  status: string;
  attemptCount: number;
  attention: string;
}

export interface DbmsActivity {
  id: string;
  timestamp: string;
  user: string;
  operation: string;
  entity: string;
  action: string;
  success: boolean;
  /** @nullable */
  query?: string | null;
}

export interface DashboardSummary {
  metrics: DashboardSummaryMetrics;
  statusDistribution: DashboardSummaryStatusDistributionItem[];
  failureReasons: DashboardSummaryFailureReasonsItem[];
  zonePerformance: DashboardSummaryZonePerformanceItem[];
  agentPerformance: DashboardSummaryAgentPerformanceItem[];
  highRiskShipments: HighRiskShipment[];
  recentActivity: DbmsActivity[];
}

export type AnalyticsOverviewFailureByReasonItem = { [key: string]: unknown };

export type AnalyticsOverviewFailureByZoneItem = { [key: string]: unknown };

export type AnalyticsOverviewFailureByAgentItem = { [key: string]: unknown };

export type AnalyticsOverviewFailureByPaymentItem = { [key: string]: unknown };

export type AnalyticsOverviewDistanceRangesItem = { [key: string]: unknown };

export type AnalyticsOverviewAttemptsPerShipmentItem = { [key: string]: unknown };

export interface AnalyticsOverview {
  failureByReason: AnalyticsOverviewFailureByReasonItem[];
  failureByZone: AnalyticsOverviewFailureByZoneItem[];
  failureByAgent: AnalyticsOverviewFailureByAgentItem[];
  failureByPayment: AnalyticsOverviewFailureByPaymentItem[];
  distanceRanges: AnalyticsOverviewDistanceRangesItem[];
  attemptsPerShipment: AnalyticsOverviewAttemptsPerShipmentItem[];
  returnRate: number;
  successRate: number;
}

export type PredictionFeatures = { [key: string]: unknown };

export interface Prediction {
  predictionId: string;
  shipmentId: string;
  failureProbability: number;
  riskLevel: string;
  predictedAt: string;
  modelVersion: string;
  features: PredictionFeatures;
}

export type DbmsOperationInputOperation = typeof DbmsOperationInputOperation[keyof typeof DbmsOperationInputOperation];


export const DbmsOperationInputOperation = {
  filter: 'filter',
  sort: 'sort',
  aggregate: 'aggregate',
  'group-analysis': 'group-analysis',
  '1-hop': '1-hop',
  '2-hop': '2-hop',
  'multi-hop': 'multi-hop',
  constraints: 'constraints',
  indexes: 'indexes',
  transaction: 'transaction',
} as const;

export interface DbmsOperationInput {
  operation: DbmsOperationInputOperation;
  /** @nullable */
  shipmentId?: string | null;
}

export type DbmsOperationResultParameters = { [key: string]: unknown };

export type DbmsOperationResultResultItem = { [key: string]: unknown };

export interface DbmsOperationResult {
  operation: string;
  query: string;
  parameters: DbmsOperationResultParameters;
  success: boolean;
  result: DbmsOperationResultResultItem[];
  affected: number;
}

export type SearchParameter = string;

export type PageParameter = number;

export type GetCustomersParams = {
search?: SearchParameter;
/**
 * @minimum 1
 */
page?: PageParameter;
};

export type GetOrdersParams = {
search?: SearchParameter;
};

export type GetShipmentsParams = {
search?: SearchParameter;
status?: string;
};