# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IntelliRoute graph logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptInput {
  status: string;
  /** @nullable */
  failureReasonId?: string | null;
  remarks: string;
}