# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IntelliRoute graph logistics API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: string;
}