# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IntelliRoute graph logistics API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryAgentPerformanceItem } from './dashboardSummaryAgentPerformanceItem';
import type { DashboardSummaryFailureReasonsItem } from './dashboardSummaryFailureReasonsItem';
import type { DashboardSummaryMetrics } from './dashboardSummaryMetrics';
import type { DashboardSummaryStatusDistributionItem } from './dashboardSummaryStatusDistributionItem';
import type { DashboardSummaryZonePerformanceItem } from './dashboardSummaryZonePerformanceItem';
import type { DbmsActivity } from './dbmsActivity';
import type { HighRiskShipment } from './highRiskShipment';

export interface DashboardSummary {
  metrics: DashboardSummaryMetrics;
  statusDistribution: DashboardSummaryStatusDistributionItem[];
  failureReasons: DashboardSummaryFailureReasonsItem[];
  zonePerformance: DashboardSummaryZonePerformanceItem[];
  agentPerformance: DashboardSummaryAgentPerformanceItem[];
  highRiskShipments: HighRiskShipment[];
  recentActivity: DbmsActivity[];
}