# lib/api-zod/src/generated/types/dbmsOperationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * IntelliRoute graph logistics API
 * OpenAPI spec version: 0.1.0
 */
import type { DbmsOperationResultParameters } from './dbmsOperationResultParameters';
import type { DbmsOperationResultResultItem } from './dbmsOperationResultResultItem';

export interface DbmsOperationResult {
  operation: string;
  query: string;
  parameters: DbmsOperationResultParameters;
  success: boolean;
  result: DbmsOperationResultResultItem[];
  affected: number;
}